Add API-MAP as '/api' to show all collections urls
The idea is stole straight from http://api.github.com/ to aid the the
use of a JSONAPI for the first time.
When you reach the /api endpoint your greeted with a map of all the
collections that the JSONAPI is aware of, not a 404 page.

Example response:
From the sample app with `sqlalchemy_jsonapi/tests/app.py`

```
-> % http --body get 0:5000/api
{
    "data": {
        "blog-comments": "http://localhost:500/api/blog-comments",
        "blog-posts": "http://localhost:5000/api/blog-posts",
        "logs": "http://localhost:5000/api/logs",
        "users": "http://localhost:5000/api/users"
    }
}
```

```#